extract main logic in lib
