+++ Cargo.toml
@@ -17,6 +17,10 @@ version = "1.1.1"
 
 edition = "2018"
 
+[[bin]]
+name = "cargo-deps"
+path = "src/main.rs"
+
 [dependencies]
 clap = "2"
 toml = "0.5"
