+++ src/main.rs
@@ -2,27 +2,11 @@
 
 #[macro_use]
 extern crate clap;
-extern crate toml;
 
-mod config;
-mod dep;
-mod error;
-mod graph;
-mod project;
-mod util;
-
-use crate::{
-    config::Config,
-    error::{CliError, CliResult},
-    project::Project,
-};
+use cargo_deps::config::Config;
+use cargo_deps::execute;
 use clap::{App, AppSettings, Arg, ArgMatches, SubCommand};
-use std::{
-    fs::File,
-    io::{self, BufWriter},
-    path::Path,
-    str::FromStr,
-};
+use std::str::FromStr;
 
 const USAGE: &str = "\
 cargo-deps writes a graph in dot format to standard output.
@@ -85,53 +69,4 @@ fn main() {
         let cfg = Config::from_matches(&m).unwrap_or_else(|e| e.exit());
         execute(cfg).map_err(|e| e.exit()).unwrap();
     }
-}
-
-fn execute(cfg: Config) -> CliResult<()> {
-    // Search through parent dirs for Cargo.toml.
-    is_cargo_toml(&cfg.manifest_path)?;
-    let manifest_path = util::find_manifest_file(&cfg.manifest_path)?;
-
-    // Cargo.lock must be in the same directory as Cargo.toml or in a parent directory.
-    let manifest = manifest_path.to_str().unwrap();
-    let lock_file = format!("{}.lock", &manifest[0..manifest.len() - 5]);
-    let lock_path = util::find_manifest_file(&lock_file)?;
-
-    // Graph the project.
-    let dot_file = cfg.dot_file.clone();
-    let project = Project::with_config(cfg)?;
-    let graph = project.graph(manifest_path, lock_path)?;
-
-    // Render the dot file.
-    match dot_file {
-        None => {
-            let o = io::stdout();
-            let mut bw = BufWriter::new(o.lock());
-            graph.render_to(&mut bw)
-        }
-        Some(file) => {
-            let o = File::create(&Path::new(&file)).expect("Failed to create file");
-            let mut bw = BufWriter::new(o);
-            graph.render_to(&mut bw)
-        }
-    }
-}
-
-// Check that the manifest file name is "Cargo.toml".
-fn is_cargo_toml(file_name: &str) -> CliResult<()> {
-    let path = Path::new(file_name);
-
-    if let Some(file_name) = path.file_name() {
-        if file_name != "Cargo.toml" {
-            return Err(CliError::Toml(
-                "The manifest-path must be a path to a Cargo.toml file".into(),
-            ));
-        }
-    } else {
-        return Err(CliError::Toml(
-            "The manifest path is not a valid file".into(),
-        ));
-    }
-
-    Ok(())
 }
